+++ src/test/java/com/codeborne/selenide/ex/ElementShouldTest.java
@@ -20,6 +20,7 @@ class ElementShouldTest implements WithAssertions {
     ElementShould elementShould = new ElementShould(driver, searchCriteria, prefix, Condition.appear, webElementMock, exception);
     String expectedString = "Element should be visible {by.name: selenide}\n" +
       "Element: '<null displayed:false></null>'\n" +
+      "Actual value: visible:false\n" +
       "Screenshot: null\n" +
       "Timeout: 0 ms.\n" +
       "Caused by: java.lang.Exception: Error message";