+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 5.2.9
 * support URLs with newlines
+* #469 improve error message of OR condition
 
 ## 5.2.8 (released 03.08.2019)
 * #961 Fix spam in logs "Failed to get attributes via JS..."