+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 5.2.9
 * support URLs with newlines
 * #469 improve error message of OR condition
+* upgrade to webdrivermanager:3.6.2
 
 ## 5.2.8 (released 03.08.2019)
 * #961 Fix spam in logs "Failed to get attributes via JS..."